scripts/crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyMySQL==1.1.0
python-dotenv==1.0.0
jsonschema==4.20.0
pyyaml==6.0.1
markdown==3.5.1
